fix: extract NumberInputAdornment from NumberInput
