+++ .changeset/smooth-toys-look.md
@@ -1,0 +1,10 @@
+---
+'@toptal/picasso': minor
+---
+
+---
+### NumberInput
+
+- Extracted NumberInputEndAdornment
+- Fixed design inconsistencies
+- Added 'large' size icons for NumberInputEndAdornment
