+++ packages/picasso/src/NumberInputEndAdornment/NumberInputEndAdornment.tsx
@@ -1,0 +1,168 @@
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
+import React, { RefObject } from 'react'
+import { makeStyles, Theme } from '@material-ui/core/styles'
+import { BaseProps, SizeType } from '@toptal/picasso-shared'
+import ButtonBase from '@material-ui/core/ButtonBase'
+import cx from 'classnames'
+
+import InputAdornment from '../InputAdornment'
+import Container from '../Container'
+import {
+  ArrowDownMinor16,
+  ArrowUpMinor16,
+  ArrowDownMinor24,
+  ArrowUpMinor24
+} from '../Icon'
+import styles from './styles'
+import isBrowser from '../utils/is-browser'
+
+export interface Props extends BaseProps {
+  /** Value of the `input` element. */
+  value?: string | number
+  /** Minimum value for the `input` element */
+  min?: number | string
+  /** Maximum value for the `input` element */
+  max?: number | string
+  /** Next value of the `input` element will be calculated based on step */
+  step?: number | string
+  /** Indicates whether component is in disabled state */
+  disabled?: boolean
+  /** Component size */
+  size?: SizeType<'small' | 'medium' | 'large'>
+  /** Ref of the input element */
+  inputRef: RefObject<HTMLInputElement>
+}
+
+const useStyles = makeStyles<Theme, Props>(styles, {
+  name: 'NumberInputEndAdornment'
+})
+
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+const nativeInputValueSetter = isBrowser()
+  ? (Object.getOwnPropertyDescriptor(
+      window.HTMLInputElement.prototype,
+      'value'
+    ) as PropertyDescriptor).set!
+  : undefined
+
+export const NumberInputEndAdornment = (props: Props) => {
+  const {
+    step = 1,
+    min = -Infinity,
+    max = Infinity,
+    value,
+    disabled,
+    size = 'medium',
+    inputRef
+  } = props
+
+  const classes = useStyles(props)
+
+  const normalizedStep = Number(step)
+  const normalizedValue = Number(value)
+  const normalizedMin = Number(min)
+  const normalizedMax = Number(max)
+
+  const fireEvent = (nextValue: number) => {
+    if (!nativeInputValueSetter) {
+      return
+    }
+
+    const input = inputRef?.current
+
+    nativeInputValueSetter.call(input, nextValue)
+
+    const event = new Event('input', {
+      bubbles: true,
+      cancelable: true
+    })
+
+    if (input) {
+      input.dispatchEvent(event)
+    }
+  }
+
+  const handleUpClick = () => {
+    if (typeof value === 'undefined') {
+      return
+    }
+
+    let nextValue = normalizedValue + normalizedStep
+
+    if (nextValue <= max) {
+      if (normalizedValue < normalizedMin + normalizedStep) {
+        nextValue = normalizedMin + normalizedStep
+      }
+
+      fireEvent(nextValue)
+    } else if (normalizedValue !== normalizedMax) {
+      nextValue = normalizedMax
+      fireEvent(nextValue)
+    }
+  }
+
+  const handleDownClick = () => {
+    if (typeof value === 'undefined') {
+      return
+    }
+
+    let nextValue = normalizedValue - normalizedStep
+
+    if (nextValue >= min) {
+      if (normalizedValue > normalizedMax - normalizedStep) {
+        nextValue = normalizedMax - normalizedStep
+      }
+
+      fireEvent(nextValue)
+    } else if (normalizedValue !== normalizedMin) {
+      nextValue = normalizedMin
+      fireEvent(nextValue)
+    }
+  }
+
+  const rootClassName = cx(
+    {
+      [classes.disabled]: disabled
+    },
+    classes[size],
+    classes.root
+  )
+
+  return (
+    <InputAdornment position='end'>
+      <Container flex direction='column' inline>
+        <ButtonBase
+          disabled={disabled}
+          classes={{
+            root: rootClassName
+          }}
+          onClick={handleUpClick}
+        >
+          {size === 'large' ? <ArrowUpMinor24 /> : <ArrowUpMinor16 />}
+        </ButtonBase>
+        <ButtonBase
+          disabled={disabled}
+          classes={{
+            root: rootClassName
+          }}
+          onClick={handleDownClick}
+        >
+          {size === 'large' ? <ArrowDownMinor24 /> : <ArrowDownMinor16 />}
+        </ButtonBase>
+      </Container>
+    </InputAdornment>
+  )
+}
+
+NumberInputEndAdornment.defaultProps = {
+  min: -Infinity,
+  max: Infinity,
+  value: 0,
+  step: 1,
+  disabled: false,
+  size: 'medium'
+}
+
+NumberInputEndAdornment.displayName = 'NumberInputEndAdornment'
+
+export default NumberInputEndAdornment
