+++ packages/picasso/src/NumberInputEndAdornment/index.ts
@@ -1,0 +1,2 @@
+export { default } from './NumberInputEndAdornment'
+export * from './NumberInputEndAdornment'
